requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
crispy-bootstrap5==2025.6
Django==5.2.8
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-environ==0.12.0
djangorestframework==3.16.1
filelock==3.20.0
future==1.0.0
gevent==25.9.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
PySocks==1.7.1
requests
requests-file==3.0.1
setuptools==80.9.0
sqlparse==0.5.3
tldextract==5.3.0
urllib3==2.5.0
wheel==0.45.1
zope.event==6.1
zope.interface==8.1
whitenoise
